refactor push template classes and builders
