+++ code/core/api/core.api
@@ -1030,23 +1030,6 @@ public final class com/adobe/marketing/mobile/services/ui/message/InAppMessageSe
 	public static fun values ()[Lcom/adobe/marketing/mobile/services/ui/message/InAppMessageSettings$MessageGesture;
 }
 
-public final class com/adobe/marketing/mobile/services/ui/notification/AEPPushPayload {
-	public static final field $stable I
-	public static final field Companion Lcom/adobe/marketing/mobile/services/ui/notification/AEPPushPayload$Companion;
-	public fun <init> (Ljava/util/Map;)V
-	public final fun getDeliveryId ()Ljava/lang/String;
-	public final fun getMessageData ()Ljava/util/Map;
-	public final fun getMessageId ()Ljava/lang/String;
-	public static final fun getNotificationCompatPriorityMap ()Ljava/util/Map;
-	public static final fun getNotificationCompatVisibilityMap ()Ljava/util/Map;
-	public final fun getTag ()Ljava/lang/String;
-}
-
-public final class com/adobe/marketing/mobile/services/ui/notification/AEPPushPayload$Companion {
-	public final fun getNotificationCompatPriorityMap ()Ljava/util/Map;
-	public final fun getNotificationCompatVisibilityMap ()Ljava/util/Map;
-}
-
 public final class com/adobe/marketing/mobile/services/ui/notification/NotificationConstructionFailedException : java/lang/Exception {
 	public static final field $stable I
 	public fun <init> (Ljava/lang/String;)V
