+++ code/core/src/phone/java/com/adobe/marketing/mobile/services/ui/notification/BasicTemplateNotificationBuilder.kt
@@ -14,467 +14,333 @@ package com.adobe.marketing.mobile.services.ui.notification
 import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
-import android.os.Build
-import android.os.Bundle
 import android.widget.RemoteViews
 import androidx.core.app.NotificationCompat
 import com.adobe.marketing.mobile.core.R
 import com.adobe.marketing.mobile.services.Log
 import com.adobe.marketing.mobile.services.ServiceProvider
-import com.adobe.marketing.mobile.util.StringUtils
+import org.json.JSONArray
+import org.json.JSONException
+import org.json.JSONObject
 
 /**
  * Object responsible for constructing a [NotificationCompat.Builder] object containing a basic push template notification.
  */
-internal object BasicTemplateNotificationBuilder{
-        private const val SELF_TAG = "BasicTemplateNotificationBuilder"
+internal object BasicTemplateNotificationBuilder : AEPPushTemplateNotificationBuilder() {
+    private const val SELF_TAG = "BasicTemplateNotificationBuilder"
+    private const val ACTION_BUTTON_CAPACITY = 3
 
-        @Throws(NotificationConstructionFailedException::class)
-        fun construct(
-            context: Context,
-            intent: Intent?,
-            pushTemplate: BasicPushTemplate?,
-            trackerActivityName: String?,
-            broadcastReceiverName: String?
-        ): NotificationCompat.Builder {
-            if (pushTemplate == null && intent == null) {
-                throw NotificationConstructionFailedException(
-                    "push template and intent are null, cannot build a notification."
-                )
+    @Throws(NotificationConstructionFailedException::class)
+    fun construct(
+        context: Context,
+        pushTemplate: BasicPushTemplate?,
+        trackerActivityName: String?,
+        broadcastReceiverName: String?
+    ): NotificationCompat.Builder {
+        if (pushTemplate == null) {
+            throw NotificationConstructionFailedException(
+                "push template is null, cannot build a basic template notification."
+            )
+        }
+        val cacheService = ServiceProvider.getInstance().cacheService
+            ?: throw NotificationConstructionFailedException("Cache service is null, basic template notification will not be constructed.")
+
+        Log.trace(
+            PushTemplateConstants.LOG_TAG,
+            SELF_TAG,
+            "Building a basic template push notification."
+        )
+        val packageName = context.packageName
+        val smallLayout = RemoteViews(packageName, R.layout.push_template_collapsed)
+        val expandedLayout = RemoteViews(packageName, R.layout.push_template_expanded)
+
+        // get push payload data
+        val imageUri = pushTemplate.imageUrl
+        val pushImage = downloadImage(cacheService, imageUri)
+        pushImage?.let {
+            expandedLayout.setImageViewBitmap(R.id.expanded_template_image, pushImage)
+        }
+        smallLayout.setTextViewText(R.id.notification_title, pushTemplate.title)
+        smallLayout.setTextViewText(R.id.notification_body, pushTemplate.body)
+        expandedLayout.setTextViewText(R.id.notification_title, pushTemplate.title)
+        expandedLayout.setTextViewText(
+            R.id.notification_body_expanded, pushTemplate.expandedBodyText
+        )
+
+        // Create the notification channel if needed
+        channelIdToUse = createChannelAndGetChannelID(
+            context,
+            pushTemplate.channelId,
+            pushTemplate.sound,
+            pushTemplate.getNotificationImportance()
+        )
+
+        // create the notification builder with the common settings applied
+        val builder = super.construct(context, pushTemplate, trackerActivityName, smallLayout, expandedLayout)
+
+        // add any action buttons defined for the notification
+        addActionButtons(
+            context,
+            trackerActivityName,
+            builder,
+            pushTemplate.actionButtonsString,
+            pushTemplate.tag,
+            pushTemplate.isNotificationSticky ?: false
+        )
+
+        // add a remind later button if we have a label and a timestamp
+        pushTemplate.remindLaterText?.let { remindLaterText ->
+            pushTemplate.remindLaterTimestamp?.let {
+                val remindIntent =
+                    createRemindPendingIntent(
+                        context,
+                        trackerActivityName,
+                        broadcastReceiverName,
+                        channelIdToUse ?: PushTemplateConstants.DEFAULT_CHANNEL_ID,
+                        pushTemplate
+                    )
+                builder.addAction(0, remindLaterText, remindIntent)
             }
-            Log.trace(
+        }
+
+        return builder
+    }
+
+    /**
+     * Adds action buttons for the notification.
+     *
+     * @param context the application [Context]
+     * @param trackerActivityName the [String] name of the activity to set in the created pending intent for tracking purposes
+     * @param builder the [NotificationCompat.Builder] to attach the action buttons
+     * @param actionButtonsString `String` a JSON string containing action buttons to attach
+     * to the notification
+     * notification
+     * @param tag `String` containing the tag to use when scheduling the notification
+     * @param stickyNotification [Boolean]  if false, remove the notification after the action
+     * button is pressed
+     */
+    fun addActionButtons(
+        context: Context,
+        trackerActivityName: String?,
+        builder: NotificationCompat.Builder,
+        actionButtonsString: String?,
+        tag: String?,
+        stickyNotification: Boolean
+    ) {
+        val actionButtons: List<AEPPushTemplate.ActionButton>? =
+            getActionButtonsFromString(actionButtonsString)
+        if (actionButtons.isNullOrEmpty()) {
+            return
+        }
+        for (eachButton in actionButtons) {
+            val pendingIntent: PendingIntent? =
+                if (eachButton.type === AEPPushTemplate.ActionType.DEEPLINK ||
+                    eachButton.type === AEPPushTemplate.ActionType.WEBURL
+                ) {
+                    BasicTemplateNotificationBuilder.createPendingIntent(
+                        context,
+                        trackerActivityName,
+                        eachButton.link,
+                        eachButton.label,
+                        tag,
+                        stickyNotification
+                    )
+                } else {
+                    BasicTemplateNotificationBuilder.createPendingIntent(
+                        context,
+                        trackerActivityName,
+                        null,
+                        eachButton.label,
+                        tag,
+                        stickyNotification
+                    )
+                }
+            builder.addAction(0, eachButton.label, pendingIntent)
+        }
+    }
+
+    private fun getActionButtonsFromString(actionButtons: String?): List<AEPPushTemplate.ActionButton>? {
+        if (actionButtons == null) {
+            Log.debug(
                 PushTemplateConstants.LOG_TAG,
-                SELF_TAG,
-                "Building a basic template push notification."
+                AEPPushTemplate.SELF_TAG,
+                "Exception in converting actionButtons json string to json object, Error :" +
+                    " actionButtons is null"
             )
-
-            return if (pushTemplate != null) {
-                construct(context, trackerActivityName, broadcastReceiverName, pushTemplate)
-            } else {
-                construct(context, trackerActivityName, broadcastReceiverName, intent!!)
-            }
+            return null
         }
-
-        @Throws(NotificationConstructionFailedException::class)
-        private fun construct(
-            context: Context,
-            trackerActivityName: String?,
-            broadcastReceiverName: String?,
-            pushTemplate: BasicPushTemplate
-        ): NotificationCompat.Builder {
-            val channelIdToUse = createChannelAndGetChannelID(
-                context,
-                pushTemplate.getChannelId(),
-                pushTemplate.getSound(),
-                pushTemplate.getNotificationImportance()
-            )
-
-            val packageName = ServiceProvider.getInstance()
-                .appContextService
-                .application
-                ?.packageName
-            val smallLayout = RemoteViews(packageName, R.layout.push_template_collapsed)
-            val expandedLayout = RemoteViews(packageName, R.layout.push_template_expanded)
-            val cacheService = ServiceProvider.getInstance().cacheService
-                ?: throw NotificationConstructionFailedException(
-                    (
-                        "Cache service is null, basic template push notification will not be" +
-                            " constructed."
-                        )
-                )
-
-            // get push payload data
-            val imageUri = pushTemplate.getImageUrl()
-            val pushImage = downloadImage(cacheService, imageUri)
-            pushImage?.let {
-                expandedLayout.setImageViewBitmap(R.id.expanded_template_image, pushImage)
+        val actionButtonList: MutableList<AEPPushTemplate.ActionButton> = ArrayList(
+            ACTION_BUTTON_CAPACITY
+        )
+        try {
+            val jsonArray = JSONArray(actionButtons)
+            for (i in 0 until jsonArray.length()) {
+                val jsonObject = jsonArray.getJSONObject(i)
+                val button = getActionButton(jsonObject) ?: continue
+                actionButtonList.add(button)
             }
-            smallLayout.setTextViewText(R.id.notification_title, pushTemplate.getTitle())
-            smallLayout.setTextViewText(R.id.notification_body, pushTemplate.getBody())
-            expandedLayout.setTextViewText(R.id.notification_title, pushTemplate.getTitle())
-            expandedLayout.setTextViewText(
-                R.id.notification_body_expanded, pushTemplate.getExpandedBodyText()
+        } catch (e: JSONException) {
+            Log.warning(
+                PushTemplateConstants.LOG_TAG,
+                AEPPushTemplate.SELF_TAG,
+                "Exception in converting actionButtons json string to json object, Error : %s",
+                e.localizedMessage
             )
-
-            // set any custom colors if needed
-            setCustomNotificationColors(
-                pushTemplate.getNotificationBackgroundColor(),
-                pushTemplate.getTitleTextColor(),
-                pushTemplate.getExpandedBodyTextColor(),
-                smallLayout,
-                expandedLayout,
-                R.id.basic_expanded_layout
-            )
-
-            // Create the notification
-            val builder = NotificationCompat.Builder(context, channelIdToUse)
-                .setTicker(pushTemplate.getTicker())
-                .setNumber(pushTemplate.getBadgeCount())
-                .setStyle(NotificationCompat.DecoratedCustomViewStyle())
-                .setCustomContentView(smallLayout)
-                .setCustomBigContentView(expandedLayout)
-
-            // small icon must be present, otherwise the notification will not be displayed.
-            setSmallIcon(
-                context, builder, pushTemplate.getSmallIcon(), pushTemplate.getSmallIconColor()
-            )
-
-            // set a large icon if one is present
-            setRemoteViewLargeIcon(pushTemplate.getLargeIcon(), smallLayout)
-            setRemoteViewLargeIcon(
-                pushTemplate.getLargeIcon(), expandedLayout
-            )
-
-            // set notification visibility
-            setVisibility(
-                builder, pushTemplate.getNotificationVisibility()
-            )
-
-            // add any action buttons defined for the notification
-            addActionButtons(
-                context,
-                trackerActivityName,
-                builder,
-                pushTemplate.getActionButtons(),
-                pushTemplate.getMessageId(),
-                pushTemplate.getDeliveryId(),
-                pushTemplate.getTag(),
-                pushTemplate.getStickyStatus() ?: false
-            )
-
-            // add a remind later button if we have a label and a timestamp
-            pushTemplate.getRemindLaterText()?.let { remindLaterText ->
-                pushTemplate.getRemindLaterTimestamp()?.let {
-                    val remindIntent =
-                        createRemindPendingIntent(
-                            context,
-                            trackerActivityName,
-                            broadcastReceiverName,
-                            channelIdToUse,
-                            pushTemplate
-                        )
-                    builder.addAction(0, remindLaterText, remindIntent)
-                }
-            }
-
-            // set custom sound, note this applies to API 25 and lower only as API 26 and up set the
-            // sound on the notification channel
-            setSound(context, builder, pushTemplate.getSound())
-            setNotificationClickAction(
-                context,
-                trackerActivityName,
-                builder,
-                pushTemplate.getMessageId(),
-                pushTemplate.getDeliveryId(),
-                pushTemplate.getActionUri(),
-                pushTemplate.getTag(),
-                pushTemplate.getStickyStatus() ?: false
-            )
-            setNotificationDeleteAction(
-                context,
-                trackerActivityName,
-                builder,
-                pushTemplate.getMessageId(),
-                pushTemplate.getDeliveryId()
-            )
-
-            // if API level is below 26 (prior to notification channels) then notification priority is
-            // set on the notification builder
-            if (Build.VERSION.SDK_INT < Build.VERSION_CODES.O) {
-                builder.setPriority(NotificationCompat.PRIORITY_HIGH)
-                    .setVibrate(LongArray(0)) // hack to enable heads up notifications as a HUD style
-                // notification requires a tone or vibration
-            }
-            return builder
+            return null
         }
+        return actionButtonList
+    }
 
-        @Throws(NotificationConstructionFailedException::class)
-        private fun construct(
-            context: Context,
-            trackerActivityName: String?,
-            broadcastReceiverName: String?,
-            intent: Intent
-        ): NotificationCompat.Builder {
-            val intentExtras = intent.extras
-                ?: throw NotificationConstructionFailedException(
-                    (
-                        "Intent extras are null, will not create a notification from the received" +
-                            " intent with action " +
-                            intent.action
-                        )
-                )
-            val cacheService = ServiceProvider.getInstance().cacheService
-                ?: throw NotificationConstructionFailedException(
-                    (
-                        "Cache service is null, basic template push notification will not be" +
-                            " constructed."
-                        )
-                )
-            val packageName = ServiceProvider.getInstance()
-                .appContextService
-                .application
-                ?.packageName
-
-            // get basic notification values from the intent extras
-            val titleText = intentExtras.getString(PushTemplateConstants.IntentKeys.TITLE_TEXT)
-            val bodyText = intentExtras.getString(PushTemplateConstants.IntentKeys.BODY_TEXT)
-            val expandedBodyText =
-                intentExtras.getString(PushTemplateConstants.IntentKeys.EXPANDED_BODY_TEXT)
-            val imageUri = intentExtras.getString(PushTemplateConstants.IntentKeys.IMAGE_URI)
-            val pushImage = downloadImage(cacheService, imageUri)
-            val smallLayout = RemoteViews(packageName, R.layout.push_template_collapsed)
-            val expandedLayout = RemoteViews(packageName, R.layout.push_template_expanded)
-            if (pushImage != null) {
-                expandedLayout.setImageViewBitmap(R.id.expanded_template_image, pushImage)
-            }
-            smallLayout.setTextViewText(R.id.notification_title, titleText)
-            smallLayout.setTextViewText(R.id.notification_body, bodyText)
-            expandedLayout.setTextViewText(R.id.notification_title, titleText)
-            expandedLayout.setTextViewText(R.id.notification_body_expanded, expandedBodyText)
-            val actionUri = intentExtras.getString(PushTemplateConstants.IntentKeys.ACTION_URI)
-            val messageId =
-                intentExtras.getString(PushTemplateConstants.IntentKeys.MESSAGE_ID) as String
-            val deliveryId =
-                intentExtras.getString(PushTemplateConstants.IntentKeys.DELIVERY_ID) as String
-            val remindLaterTimestamp =
-                intentExtras.getLong(PushTemplateConstants.IntentKeys.REMIND_TS)
-            val remindLaterText =
-                intentExtras.getString(PushTemplateConstants.IntentKeys.REMIND_LABEL)
-            val badgeCount = intentExtras.getInt(PushTemplateConstants.IntentKeys.BADGE_COUNT)
-            val visibility = intentExtras.getInt(PushTemplateConstants.IntentKeys.VISIBILITY)
-            val importance = intentExtras.getInt(PushTemplateConstants.IntentKeys.IMPORTANCE)
-            val channelId = intentExtras.getString(PushTemplateConstants.IntentKeys.CHANNEL_ID)
-            val notificationBackgroundColor = intentExtras.getString(
-                PushTemplateConstants.IntentKeys.NOTIFICATION_BACKGROUND_COLOR
-            )
-            val titleTextColor =
-                intentExtras.getString(PushTemplateConstants.IntentKeys.TITLE_TEXT_COLOR)
-            val expandedBodyTextColor =
-                intentExtras.getString(PushTemplateConstants.IntentKeys.EXPANDED_BODY_TEXT_COLOR)
-            val smallIcon = intentExtras.getString(PushTemplateConstants.IntentKeys.SMALL_ICON)
-            val smallIconColor =
-                intentExtras.getString(PushTemplateConstants.IntentKeys.SMALL_ICON_COLOR)
-            val largeIcon = intentExtras.getString(PushTemplateConstants.IntentKeys.LARGE_ICON)
-            val customSound = intentExtras.getString(PushTemplateConstants.IntentKeys.CUSTOM_SOUND)
-            val actionButtonsString =
-                intentExtras.getString(PushTemplateConstants.IntentKeys.ACTION_BUTTONS_STRING)
-            val ticker = intentExtras.getString(PushTemplateConstants.IntentKeys.TICKER)
-            val tag = intentExtras.getString(PushTemplateConstants.IntentKeys.TAG) as String
-            val sticky = intentExtras.getBoolean(PushTemplateConstants.IntentKeys.STICKY)
-            val channelIdToUse = createChannelAndGetChannelID(
-                context, channelId, customSound, importance
-            )
-
-            // set any custom colors if needed
-            setCustomNotificationColors(
-                notificationBackgroundColor,
-                titleTextColor,
-                expandedBodyTextColor,
-                smallLayout,
-                expandedLayout,
-                R.id.basic_expanded_layout
-            )
-
-            // Create the notification
-            val builder = NotificationCompat.Builder(context, channelIdToUse)
-                .setTicker(ticker)
-                .setNumber(badgeCount)
-                .setStyle(NotificationCompat.DecoratedCustomViewStyle())
-                .setCustomContentView(smallLayout)
-                .setCustomBigContentView(expandedLayout)
-
-            // small Icon must be present, otherwise the notification will not be displayed.
-            setSmallIcon(context, builder, smallIcon, smallIconColor)
-
-            // set a large icon if one is present
-            setRemoteViewLargeIcon(largeIcon, smallLayout)
-            setRemoteViewLargeIcon(largeIcon, expandedLayout)
-
-            // set notification visibility
-            setVisibility(builder, visibility)
-
-            // add any action buttons defined for the notification
-            addActionButtons(
-                context,
-                trackerActivityName,
-                builder,
-                actionButtonsString,
-                messageId,
-                deliveryId,
-                tag,
-                sticky
-            )
-
-            // add a remind later button if we have a label and a timestamp
-            if (!StringUtils.isNullOrEmpty(remindLaterText) && remindLaterTimestamp > 0) {
-                broadcastReceiverName?.let {
-                    val remindPendingIntent =
-                        createRemindPendingIntent(context, intentExtras, broadcastReceiverName)
-                    builder.addAction(0, remindLaterText, remindPendingIntent)
-                }
-            }
-
-            // set custom sound, note this applies to API 25 and lower only as API 26 and up set the
-            // sound on the notification channel
-            setSound(context, builder, customSound)
-            setNotificationClickAction(
-                context, trackerActivityName, builder, messageId, deliveryId, actionUri, tag, sticky
-            )
-
-            // set notification delete action
-            setNotificationDeleteAction(
-                context, trackerActivityName, builder, messageId, deliveryId
-            )
-
-            // if API level is below 26 (prior to notification channels) then notification priority is
-            // set on the notification builder
-            if (Build.VERSION.SDK_INT < Build.VERSION_CODES.O) {
-                builder.setPriority(NotificationCompat.PRIORITY_HIGH)
-                    .setVibrate(LongArray(0)) // hack to enable heads up notifications as a HUD style
-                // notification requires a tone or vibration
-            }
-            return builder
-        }
-
-        private fun createRemindPendingIntent(
-            context: Context,
-            trackerActivityName: String?,
-            broadcastReceiverName: String?,
-            channelId: String,
-            pushTemplate: AEPPushTemplate
-        ): PendingIntent? {
-            if (broadcastReceiverName.isNullOrEmpty()) {
+    private fun getActionButton(jsonObject: JSONObject): AEPPushTemplate.ActionButton? {
+        return try {
+            val label = jsonObject.getString(AEPPushTemplate.ActionButtons.LABEL)
+            if (label.isEmpty()) {
+                Log.debug(PushTemplateConstants.LOG_TAG, AEPPushTemplate.SELF_TAG, "Label is empty")
                 return null
             }
+            var uri: String? = null
+            val type = jsonObject.getString(AEPPushTemplate.ActionButtons.TYPE)
+            if (type == AEPPushTemplate.ActionButtonType.WEBURL || type == AEPPushTemplate.ActionButtonType.DEEPLINK) {
+                uri = jsonObject.optString(AEPPushTemplate.ActionButtons.URI)
+            }
             Log.trace(
                 PushTemplateConstants.LOG_TAG,
-                SELF_TAG,
-                "Creating a remind later pending intent from a push template object."
+                AEPPushTemplate.SELF_TAG,
+                "Creating an ActionButton with label (%s), uri (%s), and type (%s)",
+                label,
+                uri,
+                type
             )
-
-            val remindIntent = Intent(PushTemplateConstants.IntentActions.REMIND_LATER_CLICKED)
-            broadcastReceiverName.let {
-                val broadcastReceiver = Class.forName(broadcastReceiverName)
-                remindIntent.setClass(context.applicationContext, broadcastReceiver::class.java)
-            }
-
-            remindIntent.flags = Intent.FLAG_ACTIVITY_SINGLE_TOP
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.TYPE, pushTemplate.getTemplateType()?.value
-            )
-            remindIntent.putExtra(PushTemplateConstants.IntentKeys.TRACKER_NAME, trackerActivityName)
-            remindIntent.putExtra(PushTemplateConstants.IntentKeys.BROADCAST_RECEIVER_NAME, broadcastReceiverName)
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.IMAGE_URI, pushTemplate.getImageUrl()
-            )
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.ACTION_URI, pushTemplate.getActionUri()
-            )
-            remindIntent.putExtra(PushTemplateConstants.IntentKeys.CHANNEL_ID, channelId)
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.CUSTOM_SOUND, pushTemplate.getSound()
-            )
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.TITLE_TEXT,
-                pushTemplate.getTitle()
-            )
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.BODY_TEXT,
-                pushTemplate.getBody()
-            )
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.EXPANDED_BODY_TEXT,
-                pushTemplate.getExpandedBodyText()
-            )
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.NOTIFICATION_BACKGROUND_COLOR,
-                pushTemplate.getNotificationBackgroundColor()
-            )
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.TITLE_TEXT_COLOR,
-                pushTemplate.getTitleTextColor()
-            )
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.EXPANDED_BODY_TEXT_COLOR,
-                pushTemplate.getExpandedBodyTextColor()
-            )
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.MESSAGE_ID, pushTemplate.getMessageId()
-            )
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.DELIVERY_ID, pushTemplate.getDeliveryId()
-            )
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.SMALL_ICON, pushTemplate.getSmallIcon()
-            )
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.SMALL_ICON_COLOR,
-                pushTemplate.getSmallIconColor()
-            )
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.LARGE_ICON, pushTemplate.getLargeIcon()
-            )
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.VISIBILITY,
-                pushTemplate.getNotificationVisibility()
-            )
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.IMPORTANCE,
-                pushTemplate.getNotificationImportance()
-            )
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.BADGE_COUNT, pushTemplate.getBadgeCount()
-            )
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.REMIND_TS, pushTemplate.getRemindLaterTimestamp()
-            )
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.REMIND_LABEL, pushTemplate.getRemindLaterText()
-            )
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.ACTION_BUTTONS_STRING,
-                pushTemplate.getActionButtons()
-            )
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.STICKY, pushTemplate.getStickyStatus()
-            )
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.TAG, pushTemplate.getTag()
-            )
-            remindIntent.putExtra(
-                PushTemplateConstants.IntentKeys.TICKER, pushTemplate.getTicker()
-            )
-            return PendingIntent.getBroadcast(
-                context,
-                0,
-                remindIntent,
-                PendingIntent.FLAG_MUTABLE or PendingIntent.FLAG_UPDATE_CURRENT
-            )
-        }
-
-        private fun createRemindPendingIntent(
-            context: Context,
-            intentExtras: Bundle,
-            broadcastReceiverName: String?
-        ): PendingIntent {
-            Log.trace(
+            AEPPushTemplate.ActionButton(label, uri, type)
+        } catch (e: JSONException) {
+            Log.warning(
                 PushTemplateConstants.LOG_TAG,
-                SELF_TAG,
-                "Creating a remind later pending intent from an intent extras bundle."
+                AEPPushTemplate.SELF_TAG,
+                "Exception in converting actionButtons json string to json object, Error : %s",
+                e.localizedMessage
             )
-
-            val remindIntent = Intent(PushTemplateConstants.IntentActions.REMIND_LATER_CLICKED)
-            broadcastReceiverName?.let {
-                val broadcastReceiverClass = Class.forName(broadcastReceiverName)
-                remindIntent.setClass(context.applicationContext, broadcastReceiverClass::class.java)
-            }
-            remindIntent.flags = Intent.FLAG_ACTIVITY_SINGLE_TOP
-            remindIntent.putExtras(intentExtras)
-            return PendingIntent.getBroadcast(
-                context,
-                0,
-                remindIntent,
-                PendingIntent.FLAG_MUTABLE or PendingIntent.FLAG_UPDATE_CURRENT
-            )
+            null
         }
+    }
+
+    private fun createRemindPendingIntent(
+        context: Context,
+        trackerActivityName: String?,
+        broadcastReceiverName: String?,
+        channelId: String,
+        pushTemplate: BasicPushTemplate
+    ): PendingIntent? {
+        if (broadcastReceiverName.isNullOrEmpty()) {
+            return null
+        }
+        Log.trace(
+            PushTemplateConstants.LOG_TAG,
+            SELF_TAG,
+            "Creating a remind later pending intent from a push template object."
+        )
+
+        val remindIntent = Intent(PushTemplateConstants.IntentActions.REMIND_LATER_CLICKED)
+        broadcastReceiverName.let {
+            val broadcastReceiver = Class.forName(broadcastReceiverName)
+            remindIntent.setClass(context.applicationContext, broadcastReceiver::class.java)
+        }
+
+        remindIntent.flags = Intent.FLAG_ACTIVITY_SINGLE_TOP
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.TYPE, pushTemplate.templateType?.value
+        )
+        remindIntent.putExtra(PushTemplateConstants.IntentKeys.TRACKER_NAME, trackerActivityName)
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.BROADCAST_RECEIVER_NAME,
+            broadcastReceiverName
+        )
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.IMAGE_URI, pushTemplate.imageUrl
+        )
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.ACTION_URI, pushTemplate.actionUri
+        )
+        remindIntent.putExtra(PushTemplateConstants.IntentKeys.CHANNEL_ID, channelId)
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.CUSTOM_SOUND, pushTemplate.sound
+        )
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.TITLE_TEXT,
+            pushTemplate.title
+        )
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.BODY_TEXT,
+            pushTemplate.body
+        )
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.EXPANDED_BODY_TEXT,
+            pushTemplate.expandedBodyText
+        )
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.NOTIFICATION_BACKGROUND_COLOR,
+            pushTemplate.notificationBackgroundColor
+        )
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.TITLE_TEXT_COLOR,
+            pushTemplate.titleTextColor
+        )
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.EXPANDED_BODY_TEXT_COLOR,
+            pushTemplate.expandedBodyTextColor
+        )
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.SMALL_ICON, pushTemplate.smallIcon
+        )
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.SMALL_ICON_COLOR,
+            pushTemplate.smallIconColor
+        )
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.LARGE_ICON, pushTemplate.largeIcon
+        )
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.VISIBILITY,
+            pushTemplate.getNotificationVisibility()
+        )
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.IMPORTANCE,
+            pushTemplate.getNotificationImportance()
+        )
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.BADGE_COUNT, pushTemplate.badgeCount
+        )
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.REMIND_TS, pushTemplate.remindLaterTimestamp
+        )
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.REMIND_LABEL, pushTemplate.remindLaterText
+        )
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.ACTION_BUTTONS_STRING,
+            pushTemplate.actionButtonsString
+        )
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.STICKY, pushTemplate.isNotificationSticky
+        )
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.TAG, pushTemplate.tag
+        )
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.TICKER, pushTemplate.ticker
+        )
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.PAYLOAD_VERSION, pushTemplate.payloadVersion
+        )
+        remindIntent.putExtra(
+            PushTemplateConstants.IntentKeys.PRIORITY,
+            pushTemplate.notificationPriority
+        )
+
+        return PendingIntent.getBroadcast(
+            context,
+            0,
+            remindIntent,
+            PendingIntent.FLAG_MUTABLE or PendingIntent.FLAG_UPDATE_CURRENT
+        )
+    }
 }
