+++ code/core/src/phone/java/com/adobe/marketing/mobile/services/ui/notification/LegacyNotificationBuilder.kt
@@ -17,111 +17,105 @@ import androidx.core.app.NotificationCompat
 import com.adobe.marketing.mobile.services.Log
 
 /**
- * Object responsible for constructing a legacy (non [RemoteViews] based) push notification.
+ * Object responsible for constructing a legacy push notification.
  */
-internal object LegacyNotificationBuilder {
-        private const val SELF_TAG = "LegacyNotificationBuilder"
+internal object LegacyNotificationBuilder : AEPPushTemplateNotificationBuilder() {
+    private const val SELF_TAG = "LegacyNotificationBuilder"
 
-        fun construct(
-            context: Context,
-            pushTemplate: BasicPushTemplate?,
-            trackerActivityName: String?
-        ): NotificationCompat.Builder {
-            if (pushTemplate == null) {
-                throw NotificationConstructionFailedException(
-                    "push template is null, cannot build a notification."
-                )
-            }
-
-            Log.trace(
-                PushTemplateConstants.LOG_TAG,
-                SELF_TAG,
-                "Building a legacy style push notification."
+    fun construct(
+        context: Context,
+        pushTemplate: BasicPushTemplate?,
+        trackerActivityName: String?
+    ): NotificationCompat.Builder {
+        if (pushTemplate == null) {
+            throw NotificationConstructionFailedException(
+                "push template is null, cannot build a notification."
             )
-
-            // create the notification channel
-            val channelId: String = createChannelAndGetChannelID(
-                context,
-                pushTemplate.getChannelId(),
-                pushTemplate.getSound(),
-                pushTemplate.getNotificationImportance()
-            )
-
-            // Create the notification builder object and set the ticker, title, body, and badge count
-            val builder = NotificationCompat.Builder(
-                context, channelId
-            )
-                .setTicker(pushTemplate.getTicker())
-                .setContentTitle(pushTemplate.getTitle())
-                .setContentText(pushTemplate.getBody())
-                .setNumber(pushTemplate.getBadgeCount())
-
-            // set a large icon if one is present
-            setLargeIcon(
-                builder,
-                pushTemplate.getImageUrl(),
-                pushTemplate.getTitle(),
-                pushTemplate.getExpandedBodyText()
-            )
-
-            // small Icon must be present, otherwise the notification will not be displayed.
-            setSmallIcon(
-                context,
-                builder,
-                pushTemplate.getSmallIcon(),
-                pushTemplate.getSmallIconColor()
-            )
-
-            // set notification visibility
-            setVisibility(
-                builder, pushTemplate.getNotificationVisibility()
-            )
-
-            // add any action buttons defined for the notification
-            addActionButtons(
-                context,
-                trackerActivityName,
-                builder,
-                pushTemplate.getActionButtons(),
-                pushTemplate.getMessageId(),
-                pushTemplate.getDeliveryId(),
-                pushTemplate.getTag(),
-                pushTemplate.getStickyStatus() ?: false
-            )
-
-            // set custom sound, note this applies to API 25 and lower only as API 26 and up set the
-            // sound on the notification channel
-            setSound(context, builder, pushTemplate.getSound())
-
-            // assign a click action pending intent to the notification
-            setNotificationClickAction(
-                context,
-                trackerActivityName,
-                builder,
-                pushTemplate.getMessageId(),
-                pushTemplate.getDeliveryId(),
-                pushTemplate.getActionUri(),
-                pushTemplate.getTag(),
-                pushTemplate.getStickyStatus() ?: false
-            )
-
-            // set notification delete action
-            setNotificationDeleteAction(
-                context,
-                trackerActivityName,
-                builder,
-                pushTemplate.getMessageId(),
-                pushTemplate.getDeliveryId()
-            )
-
-            // if API level is below 26 (prior to notification channels) then notification priority is
-            // set on the notification builder
-            if (Build.VERSION.SDK_INT < Build.VERSION_CODES.O) {
-                builder.setPriority(NotificationCompat.PRIORITY_HIGH)
-                    .setVibrate(LongArray(0)) // hack to enable heads up notifications as a HUD style
-                // notification requires a tone or vibration
-            }
-
-            return builder
         }
+
+        Log.trace(
+            PushTemplateConstants.LOG_TAG,
+            SELF_TAG,
+            "Building a legacy style push notification."
+        )
+
+        // create the notification channel
+        val channelId: String = createChannelAndGetChannelID(
+            context,
+            pushTemplate.channelId,
+            pushTemplate.sound,
+            pushTemplate.getNotificationImportance()
+        )
+
+        // Create the notification builder object and set the ticker, title, body, and badge count
+        val builder = NotificationCompat.Builder(
+            context, channelId
+        )
+            .setTicker(pushTemplate.ticker)
+            .setContentTitle(pushTemplate.title)
+            .setContentText(pushTemplate.body)
+            .setNumber(pushTemplate.badgeCount)
+
+        // set a large icon if one is present
+        setLargeIcon(
+            builder,
+            pushTemplate.imageUrl,
+            pushTemplate.title,
+            pushTemplate.expandedBodyText
+        )
+
+        // small Icon must be present, otherwise the notification will not be displayed.
+        setSmallIcon(
+            context,
+            builder,
+            pushTemplate.smallIcon,
+            pushTemplate.smallIconColor
+        )
+
+        // set notification visibility
+        setVisibility(
+            builder, pushTemplate.getNotificationVisibility()
+        )
+
+        // add any action buttons defined for the notification
+        BasicTemplateNotificationBuilder.addActionButtons(
+            context,
+            trackerActivityName,
+            builder,
+            pushTemplate.actionButtonsString,
+            pushTemplate.tag,
+            pushTemplate.isNotificationSticky ?: false
+        )
+
+        // set custom sound, note this applies to API 25 and lower only as API 26 and up set the
+        // sound on the notification channel
+        setSound(context, builder, pushTemplate.sound)
+
+        // assign a click action pending intent to the notification
+        setNotificationClickAction(
+            context,
+            trackerActivityName,
+            builder,
+            pushTemplate.actionUri,
+            pushTemplate.tag,
+            pushTemplate.isNotificationSticky ?: false
+        )
+
+        // set notification delete action
+        setNotificationDeleteAction(
+            context,
+            trackerActivityName,
+            builder
+        )
+
+        // if API level is below 26 (prior to notification channels) then notification priority is
+        // set on the notification builder
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.O) {
+            builder.setPriority(NotificationCompat.PRIORITY_HIGH)
+                .setVibrate(LongArray(0)) // hack to enable heads up notifications as a HUD style
+            // notification requires a tone or vibration
+        }
+
+        return builder
+    }
 }
