+++ code/core/src/phone/java/com/adobe/marketing/mobile/services/ui/notification/ManualCarouselPushTemplate.kt
@@ -1,0 +1,27 @@
+/*
+  Copyright 2024 Adobe. All rights reserved.
+  This file is licensed to you under the Apache License, Version 2.0 (the "License");
+  you may not use this file except in compliance with the License. You may obtain a copy
+  of the License at http://www.apache.org/licenses/LICENSE-2.0
+  Unless required by applicable law or agreed to in writing, software distributed under
+  the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+  OF ANY KIND, either express or implied. See the License for the specific language
+  governing permissions and limitations under the License.
+*/
+
+package com.adobe.marketing.mobile.services.ui.notification
+
+import android.content.Intent
+
+internal class ManualCarouselPushTemplate : CarouselPushTemplate {
+    internal var intentAction: String? = null
+        private set
+    internal var centerImageIndex: Int? = PushTemplateConstants.DefaultValues.MANUAL_CAROUSEL_START_INDEX
+        private set
+
+    constructor(data: Map<String, String>?) : super(data)
+    constructor(intent: Intent?) : super(intent) {
+        intentAction = intent?.action
+        centerImageIndex = intent?.getIntExtra(PushTemplateConstants.IntentKeys.CENTER_IMAGE_INDEX, PushTemplateConstants.DefaultValues.MANUAL_CAROUSEL_START_INDEX)
+    }
+}
